requirements.txt
openai>=1.30.0
langchain>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langgraph>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn>=0.30.0
jupyter>=1.0.0
ipykernel>=6.29.0
nbformat>=5.10.0
langfuse>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
